Als er geen erotica filter ingesteld was, dan gaf het submitten van de user preferences een SQL error. Fixes #956

+++ lib/SpotUserSystem.php
@@ -703,7 +703,10 @@ class SpotUserSystem {
 	 */
 	function removeIndexFilter($userId) {
 		$tmpFilter = $this->_db->getUserIndexFilter($userId);
-		$this->_db->deleteFilter($userId, $tmpFilter['id'], 'index_filter');
+		
+		if (!empty($tmpFilter)) {
+			$this->_db->deleteFilter($userId, $tmpFilter['id'], 'index_filter');
+		} # if
 	} # removeIndexFilter
 
 	/*
